@@ -7,8 +7,8 @@
   "scripts": {
     "tsc": "tsc",
     "serve": "node build/server.js",
-    "serve-ts": "npx ts-node --files src/server.ts",
-    "dev": "npx nodemon --files src/server.ts",
+    "serve-ts": "npx ts-node src/server.ts",
+    "dev": "npx nodemon src/server.ts",
     "dev-old": "npx ts-node-dev --respawn --transpile-only --debug --watch src src/server.ts",
     "prod": "tsc && node ./build/app.js",
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -22,30 +22,35 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@types/body-parser": "^1.19.2",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/http-server": "^0.12.1",
     "@types/jsonwebtoken": "^8.5.8",
-    "argon2": "^0.28.5",
-    "body-parser": "^1.19.2",
+    "@types/lodash.deburr": "^4.1.7",
+    "@types/sanitize-html": "^2.6.2",
+    "argon2": "^0.28.7",
+    "body-parser": "^1.20.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.0.0",
-    "ejs": "^3.1.6",
-    "express": "^4.17.3",
+    "dotenv": "^16.0.1",
+    "ejs": "^3.1.8",
+    "express": "^4.18.1",
+    "express-validator": "^6.14.2",
     "http-status-codes": "^2.2.0",
     "jsonwebtoken": "^8.5.1",
-    "mongodb": "^4.4.1",
-    "mongoose": "^6.2.9",
-    "typescript": "^4.6.2"
+    "lodash.deburr": "^4.1.0",
+    "mongodb": "^4.8.1",
+    "mongoose": "^6.5.2",
+    "mongoose-slug-updater": "^3.3.0",
+    "sanitize-html": "^2.7.1",
+    "typescript": "^4.7.4"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
-    "eslint": "^8.14.0",
-    "eslint-import-resolver-typescript": "^2.7.1",
-    "newman": "^5.3.2",
-    "mongoose-slug-updater": "^3.3.0",
-    "nodemon": "^2.0.15",
-    "ts-node": "^10.7.0"
+    "@typescript-eslint/eslint-plugin": "^5.30.7",
+    "@typescript-eslint/parser": "^5.30.7",
+    "eslint": "^8.20.0",
+    "eslint-import-resolver-typescript": "^3.2.7",
+    "nodemon": "^2.0.19",
+    "ts-node": "^10.9.1"
   }
 }
