@@ -20,6 +20,8 @@ abstract class AbstractController {
     /** @abstract Service of a specific entity */
     abstract service:Service;
 
+    abstract name:string;
+
     /** @static UserHistory Service */
     // this is too soon. thi
     //static userHistory:UserHistory = UserHistory.getInstance();
@@ -63,6 +65,7 @@ abstract class AbstractController {
             return updatedModelResponse;
 
         LogHelper.debug("Service response from update is undefined");
+
         return ErrorResponse.create(
             new Error(ReasonPhrases.INTERNAL_SERVER_ERROR),
             StatusCodes.INTERNAL_SERVER_ERROR,
@@ -89,8 +92,9 @@ abstract class AbstractController {
      * @return {ApiResponseContract} Promise containing a list of documents
     */
     public async list(requestData:any):Promise<ApiResponseContract> {
-        LogHelper.log("Controller list : ", requestData);
+
         const query = QueryBuilder.build(requestData);
+        LogHelper.log("Controller list : ", query, requestData);
         return await this.service.all(query);
     }
 
@@ -138,10 +142,10 @@ abstract class AbstractController {
 
         const userHistoryService:UsersHistoryService = UsersHistoryService.getInstance(UserHistory.getInstance());
 
-        LogHelper.log("Create UserHistory");
+        LogHelper.log("Create UserHistory ", req.user, );
 
         //User id
-        const user:mongoose.ObjectId = req.user.user_id;
+        const user:mongoose.ObjectId = req.user.id;
 
         //IP Address
         const ipAddress = req.ip;
