@@ -2,6 +2,7 @@ import * as jwt from "jsonwebtoken";
 import config from "../../config";
 import {JwtPayload, VerifyErrors} from "jsonwebtoken";
 import {now} from "../../Helpers/DateTime";
+import LogHelper from "../../Monitoring/Helpers/LogHelper";
 
 /**
  * Controller to manage the token operation
@@ -15,6 +16,7 @@ export class TokenController {
      */
     public static generate(encapsulateData:any):string
     {
+        LogHelper.debug('Generate with the data', encapsulateData);
         //verify type of the encapsulate Data ?
         return jwt.sign(encapsulateData, config.tokenSecret, config.jwt.defaultOptions);
     }
@@ -128,10 +130,15 @@ export class TokenController {
             verifiedToken.iat >= 0 &&
             verifiedToken.exp !== undefined &&
             verifiedToken.exp >= 0 &&
-            verifiedToken.user_id !== undefined &&
+            verifiedToken.id !== undefined &&
             verifiedToken.username !== undefined &&
             verifiedToken.role !== undefined;
     }
 
+    public static generateUserToken(user:any):string
+    {
+        LogHelper.debug(user);
+        return TokenController.generate(user);
+    }
 
 }
